+++ lib/tsai_ui.dart
@@ -5,6 +5,7 @@ export 'src/components/button/tsai_button.dart';
 export 'src/components/button/tsai_button_theme.dart';
 export 'src/components/input/tsai_code_input.dart';
 export 'src/components/input/tsai_input.dart';
+export 'src/components/link/tsai_link.dart';
 export 'src/components/select/tsai_select.dart';
 export 'src/components/selection/tsai_selection_controls.dart';
 export 'src/components/typography/tsai_text.dart';